+++ packages/twenty-front/src/modules/captcha/constants/PathsRequiringCaptcha.ts
